Verder gegaan met opschonen en uitbreiden van functionaliteiten
# De lay-out is her-en-der aangepast.
# De code is op verschillende plekken geoptimaliseerd door consequent
gebruik te maken van functies.
# Her en der is commentaar geplaatst ter verduidelijking wat er gebeurd.
# De style-sheet is hernoemd en uitgebreid zodat het in Firefox en IE
werkt.
# Het is nu mogelijk huizen te zoeken op basis van een cirkel rond een
coordinaat => Hiervoor is de tabel-structuur wel aangepast.
# changeTable_24-04-2013.php past de tabelstructuur aan. Eenmaal runnen
en klaar is kees.
# checkTables.php is toegevoegd : deze zoekt huizen in de ene database
en kijkt of deze ook allemaal voorkomen in de andere database. Op deze
manier kunnen wees-huizen worden gesignaleerd.

+++ admin/delete.php
@@ -10,22 +10,36 @@ if(isset($_REQUEST['id'])) {
 	$id = $_REQUEST['id'];
 	
 	$data = getFundaData($id);
-	$deel_2 = $data['adres'];
+	$deel_2 = $data['adres'] .', '. $data['plaats'];
 	
 	if(isset($_REQUEST['zeker'])) {		
-		$sql = "DELETE FROM $TableHuizen WHERE $HuizenID like '$id'";
-		if(mysql_query($sql)) {
-			$deel_1 .= "Huis is verwijderd<br>";
-		}
+		$sql_check_unique = "SELECT * FROM $TableResultaat WHERE $ResultaatID like '$id'";
+		$result	= mysql_query($sql_check_unique);
 		
-		$sql = "DELETE FROM $TableKenmerken WHERE $KenmerkenID like '$id'";
-		if(mysql_query($sql)) {
-			$deel_1 .= "Kenmerken zijn verwijderd<br>";
-		}
-		
-		$sql = "DELETE FROM $TableResultaat WHERE $ResultaatID like '$id'";
-		if(mysql_query($sql)) {
-			$deel_1 .= "Huis uit resultaten verwijderd<br>";
+		if(mysql_num_rows($result) < 1) {
+			$sql = "DELETE FROM $TableHuizen WHERE $HuizenID like '$id'";
+			if(mysql_query($sql)) {
+				$deel_1 .= "Huis is verwijderd<br>";
+			}
+			
+			$sql = "DELETE FROM $TableKenmerken WHERE $KenmerkenID like '$id'";
+			if(mysql_query($sql)) {
+				$deel_1 .= "Kenmerken zijn verwijderd<br>";
+			}
+			
+			$sql = "DELETE FROM $TablePrijzen WHERE $PrijzenID like '$id'";
+			if(mysql_query($sql)) {
+				$deel_1 .= "Huis uit prijzen verwijderd<br>";
+			}
+			
+			$sql = "DELETE FROM $TableResultaat WHERE $ResultaatID like '$id'";
+			if(mysql_query($sql)) {
+				$deel_1 .= "Huis uit resultaten verwijderd<br>";
+			}
+			
+			$deel_1 .= "<a href='HouseDetails.php'>terug</a><br>";
+		} else {
+			$deel_1 = "Weet je het echt heel zeker? Dit huis komt namelijk in meerdere opdrachten voor<br>". $sql_check_unique ."<br><a href='?id=$id&zeker=ja&heelzeker=ja'>JA</a> | <a href='HouseDetails.php?id=$id'>NEE</a>";
 		}
 	} else {
 		$deel_1 = "Weet u zeker dat u dit huis wilt verwijderen?<br><a href='?id=$id&zeker=ja'>JA</a> | <a href='HouseDetails?id=$id'>NEE</a>";
